+++ kexec/arch/sh/crashdump-sh.c
@@ -34,8 +34,8 @@ static struct memory_range crash_memory_range[CRASH_MAX_MEMORY_RANGES];
 static int crash_sh_range_nr;
 static int crash_sh_memory_range_callback(void *UNUSED(data), int UNUSED(nr),
 					  char *str,
-					  unsigned long base,
-					  unsigned long length)
+					  unsigned long long base,
+					  unsigned long long length)
 {
 
 	struct memory_range *range = crash_memory_range;
